+++ src/bolig_ping/data_models.py
@@ -1,5 +1,7 @@
 """Data models used in the project."""
 
+import requests
+from bs4 import BeautifulSoup
 from pydantic import BaseModel
 
 
@@ -32,8 +34,6 @@ class SearchQuery(BaseModel):
             areaMin=self.min_size,
             areaMax=self.max_size,
         )
-        if self.queries:
-            arguments["text"] = ",".join(self.queries)
         url += "?" + "&".join(f"{key}={value}" for key, value in arguments.items())
         return url
 
@@ -48,6 +48,22 @@ class Flat(BaseModel):
     size: int | None
     monthly_fee: int | None
     year: int | None
+
+    @property
+    def description(self) -> str | None:
+        """Get the description of the flat.
+
+        Returns:
+            The description of the flat, or None if not available.
+        """
+        response = requests.get(url=self.url)
+        if response.ok:
+            soup = BeautifulSoup(response.content, "html.parser")
+            lines = soup.text.split("\n")
+            long_lines = [line.strip() for line in lines if len(line.strip()) > 200]
+            if long_lines:
+                return "\n".join(long_lines)
+        return None
 
     def __hash__(self) -> int:
         """Get the hash of the flat.
@@ -74,3 +90,22 @@ class Flat(BaseModel):
             ]
         )
         return html
+
+    def to_text(self) -> str:
+        """Get the flat as a text string.
+
+        Returns:
+            The flat as a text string.
+        """
+        text = "\n".join(
+            [
+                f"URL: {self.url}",
+                f"Address: {self.address}",
+                f"Price: {self.price:,} kr.",
+                f"Number of rooms: {self.num_rooms}",
+                f"Size: {self.size} m²",
+                f"Monthly fee: {self.monthly_fee:,} kr./md",
+                f"Year built: {self.year}",
+            ]
+        )
+        return text