+++ src/bolig_ping/scraper.py
@@ -112,6 +112,18 @@ def scrape_results(search_query: SearchQuery) -> list[Flat]:
         # Ensure that the progress bar is at 100% at the end
         pbar.n = pbar.total
 
+    # Filter the flats if any keyword queries were given
+    if search_query.queries:
+        flats = [
+            flat
+            for flat in tqdm(iterable=flats, desc="Filtering flats based on keywords")
+            if flat.description is not None
+            and any(
+                query.lower() in flat.description.lower()
+                for query in search_query.queries
+            )
+        ]
+
     return flats