+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
-
+### Changed
+- Changed the `--query` (`-q`) option to now search for the selected keywords in the
+  description of the flat, rather than use Boligsiden.dk's own keywords, since many
+  flats do not use these keywords. Remember that you can specify multiple queries with,
+  e.g., `bolig-ping <other-arguments> -q badekar -q altan`.
 
 
 ## [v1.2.0] - 2025-04-04